+++ c/gomlx/xlabuilder/xlabuilder.cpp
@@ -42,7 +42,7 @@ using namespace std;
 // This often lags behind Gopjrt version, if/when the C/C++ wrapper doesn't change --
 // we don't bump the version of the C/C++ code if it doesn't change.
 // But when it changes, it matches the Gopjrt version it's being released with.
-const char *GopjrtXlaBuilderVersion = "v0.6.3";
+const char *GopjrtXlaBuilderVersion = "v0.7.0";
 
 // ShapeFromXlaShape allocates and sets a new Shape struct set with the same
 // shape defined by xla::Shape. C++ only.